requirements.txt

azure-functions
azure-storage-blob
azure-data-tables
azure-ai-textanalytics
requests
opencensus-ext-azure
